updated main and relocated directory by User3198352
