--- Mergy_Hand_Pose_Estimation_Preprocessing/README.md
@@ -1,23 +1,0 @@
-### Title : Hand Pose Estimation Preprocessing
-
-- mergy by image processing task and Hand Pose Estimation
-- Two module Python codes exist in this file. One code has Hand Pose Estimation and the other has an image processing module.
-- And there is one main Python code, hand_pose_estimation.py, which aggregates them.
-- We have put the environment settings in the requirements.txt file to run this code.
-
-### Enviroments
-
-```bsh
-pip install -r requirments.txt
-```
-
-### How to run this code
-```
-python ./hand_pose_estimation.py
-```
-
-- This code is still being modified.
-
-- Developer simpleis6est, user3198352
-
-- If you have any questions about this code, leave a question on the ID above
