--- Mergy_Hand_Pose_Estimation_Preprocessing/models/label_schema.json
@@ -1,258 +1,0 @@
-{
-    "label_tree": {
-        "type": "tree",
-        "directed": true,
-        "nodes": [
-            "0",
-            "1",
-            "2",
-            "3",
-            "4",
-            "5",
-            "6",
-            "7",
-            "8",
-            "9",
-            "10",
-            "11",
-            "12",
-            "13"
-        ],
-        "edges": []
-    },
-    "label_groups": [
-        {
-            "_id": "655f0aea04ec1b64ad192818",
-            "name": "labels",
-            "label_ids": [
-                "0",
-                "1",
-                "2",
-                "3",
-                "4",
-                "5",
-                "6",
-                "7",
-                "8",
-                "9",
-                "10",
-                "11",
-                "12",
-                "13"
-            ],
-            "relation_type": "EXCLUSIVE"
-        }
-    ],
-    "all_labels": {
-        "0": {
-            "_id": "0",
-            "name": "+",
-            "color": {
-                "red": 251,
-                "green": 244,
-                "blue": 65,
-                "alpha": 255
-            },
-            "hotkey": "",
-            "domain": "CLASSIFICATION",
-            "creation_date": "2023-11-23T08:18:49.779000",
-            "is_empty": false,
-            "is_anomalous": false
-        },
-        "1": {
-            "_id": "1",
-            "name": "-",
-            "color": {
-                "red": 77,
-                "green": 226,
-                "blue": 138,
-                "alpha": 255
-            },
-            "hotkey": "",
-            "domain": "CLASSIFICATION",
-            "creation_date": "2023-11-23T08:18:49.779000",
-            "is_empty": false,
-            "is_anomalous": false
-        },
-        "2": {
-            "_id": "2",
-            "name": "0",
-            "color": {
-                "red": 99,
-                "green": 158,
-                "blue": 115,
-                "alpha": 255
-            },
-            "hotkey": "",
-            "domain": "CLASSIFICATION",
-            "creation_date": "2023-11-23T08:18:49.779000",
-            "is_empty": false,
-            "is_anomalous": false
-        },
-        "3": {
-            "_id": "3",
-            "name": "1",
-            "color": {
-                "red": 73,
-                "green": 133,
-                "blue": 149,
-                "alpha": 255
-            },
-            "hotkey": "",
-            "domain": "CLASSIFICATION",
-            "creation_date": "2023-11-23T08:18:49.779000",
-            "is_empty": false,
-            "is_anomalous": false
-        },
-        "4": {
-            "_id": "4",
-            "name": "2",
-            "color": {
-                "red": 249,
-                "green": 6,
-                "blue": 175,
-                "alpha": 255
-            },
-            "hotkey": "",
-            "domain": "CLASSIFICATION",
-            "creation_date": "2023-11-23T08:18:49.779000",
-            "is_empty": false,
-            "is_anomalous": false
-        },
-        "5": {
-            "_id": "5",
-            "name": "3",
-            "color": {
-                "red": 180,
-                "green": 83,
-                "blue": 41,
-                "alpha": 255
-            },
-            "hotkey": "",
-            "domain": "CLASSIFICATION",
-            "creation_date": "2023-11-23T08:18:49.779000",
-            "is_empty": false,
-            "is_anomalous": false
-        },
-        "6": {
-            "_id": "6",
-            "name": "4",
-            "color": {
-                "red": 61,
-                "green": 98,
-                "blue": 32,
-                "alpha": 255
-            },
-            "hotkey": "",
-            "domain": "CLASSIFICATION",
-            "creation_date": "2023-11-23T08:18:49.779000",
-            "is_empty": false,
-            "is_anomalous": false
-        },
-        "7": {
-            "_id": "7",
-            "name": "5",
-            "color": {
-                "red": 182,
-                "green": 194,
-                "blue": 237,
-                "alpha": 255
-            },
-            "hotkey": "",
-            "domain": "CLASSIFICATION",
-            "creation_date": "2023-11-23T08:18:49.779000",
-            "is_empty": false,
-            "is_anomalous": false
-        },
-        "8": {
-            "_id": "8",
-            "name": "6",
-            "color": {
-                "red": 230,
-                "green": 98,
-                "blue": 134,
-                "alpha": 255
-            },
-            "hotkey": "",
-            "domain": "CLASSIFICATION",
-            "creation_date": "2023-11-23T08:18:49.779000",
-            "is_empty": false,
-            "is_anomalous": false
-        },
-        "9": {
-            "_id": "9",
-            "name": "7",
-            "color": {
-                "red": 81,
-                "green": 61,
-                "blue": 42,
-                "alpha": 255
-            },
-            "hotkey": "",
-            "domain": "CLASSIFICATION",
-            "creation_date": "2023-11-23T08:18:49.779000",
-            "is_empty": false,
-            "is_anomalous": false
-        },
-        "10": {
-            "_id": "10",
-            "name": "8",
-            "color": {
-                "red": 175,
-                "green": 98,
-                "blue": 168,
-                "alpha": 255
-            },
-            "hotkey": "",
-            "domain": "CLASSIFICATION",
-            "creation_date": "2023-11-23T08:18:49.779000",
-            "is_empty": false,
-            "is_anomalous": false
-        },
-        "11": {
-            "_id": "11",
-            "name": "9",
-            "color": {
-                "red": 128,
-                "green": 147,
-                "blue": 192,
-                "alpha": 255
-            },
-            "hotkey": "",
-            "domain": "CLASSIFICATION",
-            "creation_date": "2023-11-23T08:18:49.779000",
-            "is_empty": false,
-            "is_anomalous": false
-        },
-        "12": {
-            "_id": "12",
-            "name": "\u00d7",
-            "color": {
-                "red": 84,
-                "green": 244,
-                "blue": 173,
-                "alpha": 255
-            },
-            "hotkey": "",
-            "domain": "CLASSIFICATION",
-            "creation_date": "2023-11-23T08:18:49.779000",
-            "is_empty": false,
-            "is_anomalous": false
-        },
-        "13": {
-            "_id": "13",
-            "name": "\u00f7",
-            "color": {
-                "red": 44,
-                "green": 185,
-                "blue": 179,
-                "alpha": 255
-            },
-            "hotkey": "",
-            "domain": "CLASSIFICATION",
-            "creation_date": "2023-11-23T08:18:49.779000",
-            "is_empty": false,
-            "is_anomalous": false
-        }
-    }
-}
