--- Mergy_Hand_Pose_Estimation_Preprocessing/requirements.txt
@@ -1,60 +1,0 @@
-absl-py==2.0.0
-attrs==23.1.0
-certifi==2023.7.22
-cffi==1.16.0
-charset-normalizer==3.3.2
-contourpy==1.1.1
-cryptography==41.0.5
-cycler==0.12.1
-docutils==0.20.1
-et-xmlfile==1.1.0
-flatbuffers==23.5.26
-fonttools==4.44.3
-idna==3.4
-importlib-metadata==6.8.0
-importlib-resources==6.1.1
-jaraco.classes==3.3.0
-jeepney==0.8.0
-keyring==24.3.0
-kiwisolver==1.4.5
-markdown-it-py==3.0.0
-matplot==0.1.9
-matplotlib==3.7.3
-mdurl==0.1.2
-mediapipe==0.10.8
-more-itertools==10.1.0
-nh3==0.2.14
-numpy==1.24.4
-opencv-contrib-python==4.8.1.78
-opencv-python==4.8.1.78
-openpyxl==3.1.2
-packaging==23.2
-pandas==2.0.3
-Pillow==10.1.0
-pkginfo==1.9.6
-protobuf==3.20.3
-pycparser==2.21
-Pygments==2.16.1
-pyloco==0.0.139
-pyparsing==3.1.1
-python-dateutil==2.8.2
-pytz==2023.3.post1
-readme-renderer==42.0
-requests==2.31.0
-requests-toolbelt==1.0.0
-rfc3986==2.0.0
-rich==13.7.0
-SecretStorage==3.3.3
-SimpleWebSocketServer==0.1.2
-six==1.16.0
-sounddevice==0.4.6
-tk==0.1.0
-twine==4.0.2
-typing==3.7.4.3
-typing_extensions==4.8.0
-tzdata==2023.3
-urllib3==2.1.0
-ushlex==0.99.1
-websocket-client==1.6.4
-zipp==3.17.0
-
