+++ model/label_schema.json
@@ -1,0 +1,258 @@
+{
+    "label_tree": {
+        "type": "tree",
+        "directed": true,
+        "nodes": [
+            "0",
+            "1",
+            "2",
+            "3",
+            "4",
+            "5",
+            "6",
+            "7",
+            "8",
+            "9",
+            "10",
+            "11",
+            "12",
+            "13"
+        ],
+        "edges": []
+    },
+    "label_groups": [
+        {
+            "_id": "655fdd5e0d7553dc736225df",
+            "name": "labels",
+            "label_ids": [
+                "0",
+                "1",
+                "2",
+                "3",
+                "4",
+                "5",
+                "6",
+                "7",
+                "8",
+                "9",
+                "10",
+                "11",
+                "12",
+                "13"
+            ],
+            "relation_type": "EXCLUSIVE"
+        }
+    ],
+    "all_labels": {
+        "0": {
+            "_id": "0",
+            "name": "+",
+            "color": {
+                "red": 55,
+                "green": 63,
+                "blue": 72,
+                "alpha": 255
+            },
+            "hotkey": "",
+            "domain": "CLASSIFICATION",
+            "creation_date": "2023-11-23T23:16:46.128000",
+            "is_empty": false,
+            "is_anomalous": false
+        },
+        "1": {
+            "_id": "1",
+            "name": "-",
+            "color": {
+                "red": 130,
+                "green": 77,
+                "blue": 4,
+                "alpha": 255
+            },
+            "hotkey": "",
+            "domain": "CLASSIFICATION",
+            "creation_date": "2023-11-23T23:16:46.128000",
+            "is_empty": false,
+            "is_anomalous": false
+        },
+        "2": {
+            "_id": "2",
+            "name": "0",
+            "color": {
+                "red": 108,
+                "green": 34,
+                "blue": 221,
+                "alpha": 255
+            },
+            "hotkey": "",
+            "domain": "CLASSIFICATION",
+            "creation_date": "2023-11-23T23:16:46.128000",
+            "is_empty": false,
+            "is_anomalous": false
+        },
+        "3": {
+            "_id": "3",
+            "name": "1",
+            "color": {
+                "red": 154,
+                "green": 201,
+                "blue": 233,
+                "alpha": 255
+            },
+            "hotkey": "",
+            "domain": "CLASSIFICATION",
+            "creation_date": "2023-11-23T23:16:46.128000",
+            "is_empty": false,
+            "is_anomalous": false
+        },
+        "4": {
+            "_id": "4",
+            "name": "2",
+            "color": {
+                "red": 115,
+                "green": 117,
+                "blue": 158,
+                "alpha": 255
+            },
+            "hotkey": "",
+            "domain": "CLASSIFICATION",
+            "creation_date": "2023-11-23T23:16:46.128000",
+            "is_empty": false,
+            "is_anomalous": false
+        },
+        "5": {
+            "_id": "5",
+            "name": "3",
+            "color": {
+                "red": 44,
+                "green": 53,
+                "blue": 126,
+                "alpha": 255
+            },
+            "hotkey": "",
+            "domain": "CLASSIFICATION",
+            "creation_date": "2023-11-23T23:16:46.128000",
+            "is_empty": false,
+            "is_anomalous": false
+        },
+        "6": {
+            "_id": "6",
+            "name": "4",
+            "color": {
+                "red": 157,
+                "green": 0,
+                "blue": 95,
+                "alpha": 255
+            },
+            "hotkey": "",
+            "domain": "CLASSIFICATION",
+            "creation_date": "2023-11-23T23:16:46.128000",
+            "is_empty": false,
+            "is_anomalous": false
+        },
+        "7": {
+            "_id": "7",
+            "name": "5",
+            "color": {
+                "red": 3,
+                "green": 62,
+                "blue": 191,
+                "alpha": 255
+            },
+            "hotkey": "",
+            "domain": "CLASSIFICATION",
+            "creation_date": "2023-11-23T23:16:46.128000",
+            "is_empty": false,
+            "is_anomalous": false
+        },
+        "8": {
+            "_id": "8",
+            "name": "6",
+            "color": {
+                "red": 230,
+                "green": 141,
+                "blue": 201,
+                "alpha": 255
+            },
+            "hotkey": "",
+            "domain": "CLASSIFICATION",
+            "creation_date": "2023-11-23T23:16:46.128000",
+            "is_empty": false,
+            "is_anomalous": false
+        },
+        "9": {
+            "_id": "9",
+            "name": "7",
+            "color": {
+                "red": 228,
+                "green": 117,
+                "blue": 36,
+                "alpha": 255
+            },
+            "hotkey": "",
+            "domain": "CLASSIFICATION",
+            "creation_date": "2023-11-23T23:16:46.128000",
+            "is_empty": false,
+            "is_anomalous": false
+        },
+        "10": {
+            "_id": "10",
+            "name": "8",
+            "color": {
+                "red": 223,
+                "green": 132,
+                "blue": 184,
+                "alpha": 255
+            },
+            "hotkey": "",
+            "domain": "CLASSIFICATION",
+            "creation_date": "2023-11-23T23:16:46.128000",
+            "is_empty": false,
+            "is_anomalous": false
+        },
+        "11": {
+            "_id": "11",
+            "name": "9",
+            "color": {
+                "red": 20,
+                "green": 137,
+                "blue": 46,
+                "alpha": 255
+            },
+            "hotkey": "",
+            "domain": "CLASSIFICATION",
+            "creation_date": "2023-11-23T23:16:46.128000",
+            "is_empty": false,
+            "is_anomalous": false
+        },
+        "12": {
+            "_id": "12",
+            "name": "*",
+            "color": {
+                "red": 105,
+                "green": 127,
+                "blue": 204,
+                "alpha": 255
+            },
+            "hotkey": "",
+            "domain": "CLASSIFICATION",
+            "creation_date": "2023-11-23T23:16:46.128000",
+            "is_empty": false,
+            "is_anomalous": false
+        },
+        "13": {
+            "_id": "13",
+            "name": "/",
+            "color": {
+                "red": 205,
+                "green": 187,
+                "blue": 181,
+                "alpha": 255
+            },
+            "hotkey": "",
+            "domain": "CLASSIFICATION",
+            "creation_date": "2023-11-23T23:16:46.128000",
+            "is_empty": false,
+            "is_anomalous": false
+        }
+    }
+}
